add scale for weight

+++ app/src/main/java/is/hello/sense/api/model/Account.java
@@ -34,7 +34,7 @@ public class Account extends ApiResponse implements Cloneable {
     private Integer height;
 
     @JsonProperty("weight")
-    private Long weight;
+    private Integer weight;
 
     @JsonProperty("dob")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = ApiService.DATE_FORMAT)
@@ -114,11 +114,11 @@ public class Account extends ApiResponse implements Cloneable {
         this.height = height;
     }
 
-    public Long getWeight() {
+    public Integer getWeight() {
         return weight;
     }
 
-    public void setWeight(Long weight) {
+    public void setWeight(int weight) {
         this.weight = weight;
     }
